perception/requirements.txt
huggingface-hub
numpy<2.0.0
transformers
torch
opencv-python-headless
pyzmq
